change csr addr type.

+++ libs/rvvm/src/rv32_csr_base.cc
@@ -9,17 +9,17 @@ rv32_csr_base::rv32_csr_base()
 rv32_csr_base::~rv32_csr_base()
 {}
 
-bool rv32_csr_base::had_addr(rv32_core *core, uint32_t addr)
+bool rv32_csr_base::had_addr(rv32_core *core, uint16_t addr)
 {
     return false;
 }
 
-bool rv32_csr_base::rd(rv32_core *core, uint32_t addr, uint32_t &out)
+bool rv32_csr_base::rd(rv32_core *core, uint16_t addr, uint32_t &out)
 {
     return false;
 }
 
-bool rv32_csr_base::wr(rv32_core *core, uint32_t addr, uint32_t in)
+bool rv32_csr_base::wr(rv32_core *core, uint16_t addr, uint32_t in)
 {
     return false;
 }
